requirements.txt

boto3==1.35.92 
django-classified==1.1.1
django-environ==0.11.2
django==5.1.8
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-memcached==1.62
social-auth-app-django==5.4.2
whitenoise==6.8.2
django-fsm==3.0.0


django-crispy-forms==2.1
crispy-bootstrap4==2024.1
django-widget-tweaks==1.5.0
django-extensions==3.2.3


django-notifications-hq==1.8.0  

stripe==7.0.0